move ligandmpnn to logits ino a method
