+++ ferritin-examples/examples/ligandmpnn-onnx/Readme.md
@@ -1,8 +1,7 @@
-# ESM2-ONNX
+# LigandMPNN-ONNX
 
 Convert ESM2_6T_35M to the ONNX format. Run it locally via RUST /  [ORT](https://ort.pyke.io)
 
 ```sh
-cargo run --example ligandmpnn-onnx -- --model-id 8M --protein-string \
-MAFSAEDVLKEYDRR\<mask\>RMEALLLSLYYPNDRKLLDYKEWSPPRVQVECPKAPVEWNNPPSEKGLIVGHFSGIKYKGEKAQASEVDVNKMCCWVSKFKDAMRRYQGIQTCKIPGKVLSDLDAKIKAYNLTVEGVEGFVRYSRVTKQHVAAFLKELRHSKQYENVNLIHYILTDKRVDIQHLEKDLVKDFKALVESAHRMRQGHMINVKYILYQLLKKHGHGPDGPDILTVKTGSKGVLYDDSFRKIYTDLGWKFTPL
+cargo run --example ligandmpnn-onnx
 ```
